模块开发--项目管理  －－－－－　办证推进标段管理和形象进度标段管理

+++ serverside/java/src/main/java/com/justonetech/biz/controller/project/ProjBidController.java
@@ -29,14 +29,16 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
+import java.util.ArrayList;
+import java.util.List;
 import java.util.Set;
 
 
 /**
  * note:项目标段管理(包括办证推进和形象进度2种标段)
  * author: system
- * create date:
- * modify date:
+ * create date: 2015-05-04
+ * modify date: 2015-07-23
  */
 @Controller
 public class ProjBidController extends BaseCRUDActionController<ProjBid> {
@@ -72,7 +74,6 @@ public class ProjBidController extends BaseCRUDActionController<ProjBid> {
         model.addAttribute("typeName", ProjBidType.getNameByCode(typeCode));
         model.addAttribute("TYPE_STAGE", ProjBidType.TYPE_STAGE.getCode());
         model.addAttribute("TYPE_NODE", ProjBidType.TYPE_NODE.getCode());
-
         return "view/project/projBid/projGrid";
     }
 
@@ -89,19 +90,14 @@ public class ProjBidController extends BaseCRUDActionController<ProjBid> {
     public void projGridDataCustom(HttpServletResponse response, String filters, String columns, int page, int rows, String typeCode, HttpSession session) {
         try {
             Page pageModel = new Page(page, rows, true);
-            String hql = "from ProjInfo where 1=1";
+            String hql = "from ProjInfo where 1 = 1 ";
             //增加项目过滤
             hql += projectRelateManager.getRelateProjectHql("id");
-
             hql += "order by no asc,id asc";
-
-            //执行查询
             QueryTranslateJq queryTranslate = new QueryTranslateJq(hql, filters);
             String query = queryTranslate.toString();
             session.setAttribute(Constants.GRID_SQL_KEY, query);
             pageModel = projInfoService.findByPage(pageModel, query);
-
-            //输出显示
             String json = GridJq.toJSON(columns, pageModel);
             sendJSON(response, json);
 
@@ -127,6 +123,7 @@ public class ProjBidController extends BaseCRUDActionController<ProjBid> {
             model.addAttribute("canEdit", sysUserManager.hasPrivilege(PrivilegeCode.PROJ_BID_NODE_EDIT));
         }
         model.addAttribute("typeCode", typeCode);
+        model.addAttribute("TYPE_STAGE", ProjBidType.TYPE_STAGE.getCode());
         model.addAttribute("projId", projId);
         model.addAttribute("projName", projInfoService.get(projId).getName());
 
@@ -152,17 +149,12 @@ public class ProjBidController extends BaseCRUDActionController<ProjBid> {
                 hql += " and typeCode='" + typeCode + "'";
             }
             hql += " order by no asc,id asc";
-
-            //执行查询
             QueryTranslateJq queryTranslate = new QueryTranslateJq(hql, filters);
             String query = queryTranslate.toString();
             session.setAttribute(Constants.GRID_SQL_KEY, query);
             pageModel = projBidService.findByPage(pageModel, query);
-
-            //输出显示
             String json = GridJq.toJSON(columns, pageModel);
             sendJSON(response, json);
-
         } catch (Exception e) {
             log.error("error", e);
             super.processException(response, e);
@@ -177,14 +169,12 @@ public class ProjBidController extends BaseCRUDActionController<ProjBid> {
      */
     @RequestMapping
     public String add(Model model, Long projId, String typeCode) {
+        String forwardUrl = modify(model, null);
         ProjBid projBid = new ProjBid();
-
-        //如需增加其他默认值请在此添加
         projBid.setProject(projInfoService.get(projId));
         projBid.setTypeCode(typeCode);
         model.addAttribute("bean", projBid);
-
-        return "view/project/projBid/input";
+        return forwardUrl;
     }
 
     /**
@@ -196,11 +186,24 @@ public class ProjBidController extends BaseCRUDActionController<ProjBid> {
      */
     @RequestMapping
     public String modify(Model model, Long id) {
-        ProjBid projBid = projBidService.get(id);
-
+        ProjBid projBid = id != null ? projBidService.get(id) : new ProjBid();
+        List<String[]> relatedUnits = new ArrayList<String[]>();
+        String[] unitContactInfos = new String[]{"建设单位", "jsdw", projBid.getJsdwName(), projBid.getJsdwPerson(), projBid.getJsdwTel(),
+                "勘察单位", "kcdw", projBid.getKcdwName(), projBid.getKcdwPerson(), projBid.getKcdwTel(),
+                "设计单位", "sjdw", projBid.getSjdwName(), projBid.getSjdwPerson(), projBid.getSjdwTel(),
+                "施工单位", "sgdw", projBid.getSgdwName(), projBid.getSgdwPerson(), projBid.getSgdwTel(),
+                "监理单位", "jldw", projBid.getJldwName(), projBid.getJldwPerson(), projBid.getJldwTel()
+        };
+        String[] oneUnitInfo;
+        for (int i = 0; i < unitContactInfos.length; i++) {
+            if (i % 5 == 0) {
+                oneUnitInfo = new String[]{unitContactInfos[i], unitContactInfos[i + 1], unitContactInfos[i + 2], unitContactInfos[i + 3], unitContactInfos[i + 4]};
+                relatedUnits.add(oneUnitInfo);
+            }
+        }
         model.addAttribute("bean", projBid);
         model.addAttribute("areas", projBid.getBelongAreaNames());
-
+        model.addAttribute("relatedUnits", relatedUnits);
         return "view/project/projBid/input";
     }
 
@@ -213,11 +216,7 @@ public class ProjBidController extends BaseCRUDActionController<ProjBid> {
      */
     @RequestMapping
     public String view(Model model, Long id) {
-        ProjBid projBid = projBidService.get(id);
-
-        model.addAttribute("bean", projBid);
-        model.addAttribute("areas", projBid.getBelongAreaNames());
-        return "view/project/projBid/view";
+        return modify(model, id).replace("input", "view");
     }
 
     /**
@@ -250,12 +249,23 @@ public class ProjBidController extends BaseCRUDActionController<ProjBid> {
             System.out.println("target.getNo() = " + target.getNo());
             String[] areaIds = request.getParameterValues("ProjBelongArea");
             String projInfoId = request.getParameter("projInfoId");
-
             target.setProject(projInfoService.get(Long.valueOf(projInfoId)));
+            //保存参建单位各信息（按顺序为： 建设单位  勘察单位  设计单位  施工单位  监理单位）
+            //参建单位名称、联系人、联系电话信息 | 参建单位类型  | 参建单位字段属性
+            String[][] relatedUnitInfos = new String[][] {StringHelper.stringToStringArray(request.getParameter("relatedUnitInfos"), ","), {"Jsdw", "Kcdw", "Sjdw", "Sgdw", "Jldw"}, {"Name", "Person", "Tel"}};
+            String grpUnitType = "";
+            for (int i = 0; i < relatedUnitInfos[0].length; i++) {
+                if (i % 3 == 0) {
+                    grpUnitType = relatedUnitInfos[1][i / 3];
+                }
+                target.getClass().getMethod("set".concat(grpUnitType).concat(relatedUnitInfos[2][i % 3]), new Class<?>[]{String.class}).invoke(target, new Object[]{relatedUnitInfos[0][i].replace("empty", "")});
+            }
             projBidService.save(target);
             //保存区县前删除保存过的信息
-            for (ProjBidArea projBidArea : target.getProjBidAreas()) {
-                projBidAreaService.delete(projBidArea);
+            if (target.getProjBidAreas() != null) {
+                for (ProjBidArea projBidArea : target.getProjBidAreas()) {
+                    projBidAreaService.delete(projBidArea);
+                }
             }
             if (areaIds != null && areaIds.length > 0) {
                 for (String areaId : areaIds) {
@@ -268,7 +278,6 @@ public class ProjBidController extends BaseCRUDActionController<ProjBid> {
                     }
                 }
             }
-
         } catch (Exception e) {
             log.error("error", e);
             super.processException(response, e);
@@ -292,8 +301,62 @@ public class ProjBidController extends BaseCRUDActionController<ProjBid> {
             projBidAreaService.delete(projBidArea);
         }
         projBidService.delete(id);
-
         sendSuccessJSON(response, "删除成功");
     }
 
+    /**
+     * 复制项目办证标段到形象进度标段对应项目下
+     *
+     * @param response
+     * @param prjBidId
+     * @throws Exception
+     */
+    @RequestMapping
+    public void duplicatePrjBid4Schedule(HttpServletResponse response, Long prjBidId) throws Exception {
+        try {
+            ProjBid projBid2BeDupSrc = projBidService.get(prjBidId);
+            ProjBid projBid2BeDupTar = new ProjBid();
+            ReflectionUtils.copyBean(projBid2BeDupSrc, projBid2BeDupTar, new String[]{
+                    "code",
+                    "name",
+                    "buildMileage",
+                    "projLink",
+                    "linkTel",
+                    "startDate",
+                    "jsdwName",
+                    "jsdwTel",
+                    "sjdwPerson",
+                    "sgdwName",
+                    "jldwName",
+                    "sjdwName",
+                    "jldwPerson",
+                    "no",
+                    "kcdwTel",
+                    "kcdwName",
+                    "jsdwPerson",
+                    "sgdwPerson",
+                    "sgdwTel",
+                    "name",
+                    "linkTel",
+                    "jldwTel",
+                    "kcdwPerson",
+                    "sjdwTel"
+            });
+            projBid2BeDupTar.setProject(projBid2BeDupSrc.getProject());
+            projBid2BeDupTar.setTypeCode(ProjBidType.TYPE_NODE.getCode());
+            projBidService.save(projBid2BeDupTar);
+            for (ProjBidArea pba : projBid2BeDupSrc.getProjBidAreas()) {
+                ProjBidArea prjBa = new ProjBidArea();
+                prjBa.setBelongArea(pba.getBelongArea());
+                prjBa.setBid(projBid2BeDupTar);
+                projBidAreaService.save(prjBa);
+            }
+        } catch (Exception e) {
+            logger.error(e.getMessage(), e);
+            super.processException(response, e);
+            sendFailureJSON(response, "复制失败");
+            return;
+        }
+        sendSuccessJSON(response, "复制成功！");
+    }
 }
